% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0aux.R
\name{is_instrumentation_key}
\alias{is_instrumentation_key}
\title{Check if string matches pattern for an instrumentation key.}
\usage{
is_instrumentation_key(x)
}
\arguments{
\item{x}{A string containing nothing else but an instrumentation key.}
}
\value{
Logical value.
}
\description{
Check if string matches pattern for an instrumentation key.
}
