% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_snippet.R
\name{trackEvent}
\alias{trackEvent}
\title{Sends an event to Application Insights}
\usage{
trackEvent(session, name, properties)
}
\arguments{
\item{session}{The \code{session} object passed to function given to \code{shinyServer}.}

\item{name}{Name of the event.}

\item{properties}{List of properties to track. \code{appId} is automatically inserted.}
}
\value{
Method sends data to client's browser; returns the sent list, invisibly.
}
\description{
Sends an event to Application Insights
}
